+++ CHANGES.md
@@ -8,7 +8,12 @@ To be released.
 
  -  (Libplanet.Store) Changed `IStateStore.Commit()` to return an `ITrie`
     with either its `Root` as `null` or a `HashNode`.  [[#3610]]
+ -  (Libplanet.Store) Removed `IStateStore.PruneStates()` method.
+    [[#3613], [#3614]]
 
+[#3610]: https://github.com/planetarium/libplanet/pull/3610
+[#3613]: https://github.com/planetarium/libplanet/issues/3613
+[#3614]: https://github.com/planetarium/libplanet/pull/3614
 
 Version 3.9.4
 -------------